+++ src/manifolds/Circle.jl
@@ -9,7 +9,7 @@ real-valued points in ``[-π,π)`` or complex-valued points ``z ∈ ℂ`` of abs
     Circle(𝔽=ℝ)
 
 Generate the `ℝ`-valued Circle represented by angles, which
-alternatively can be set to use the [`AbstractNumbers`](https://juliamanifolds.github.io/ManifoldsBase.jl/stable/types.html#number-system) `𝔽=ℂ` to obtain the circle
+alternatively can be set to use the [`AbstractNumbers`](@extref ManifoldsBase number-system) `𝔽=ℂ` to obtain the circle
 represented by `ℂ`-valued circle of unit numbers.
 """
 struct Circle{𝔽} <: AbstractManifold{𝔽} end