+++ src/manifolds/Rotations.jl
@@ -157,7 +157,7 @@ end
 @doc raw"""
     injectivity_radius(M::Rotations, ::PolarRetraction)
 
-Return the radius of injectivity for the [`PolarRetraction`](https://juliamanifolds.github.io/ManifoldsBase.jl/stable/retractions.html#ManifoldsBase.PolarRetraction) on the
+Return the radius of injectivity for the [`PolarRetraction`](@extref `ManifoldsBase.PolarRetraction`) on the
 [`Rotations`](@ref) `M` which is ``\frac{π}{\sqrt{2}}``.
 """
 injectivity_radius(::Rotations, ::PolarRetraction)
@@ -172,7 +172,7 @@ end
 Compute a vector from the tangent space ``T_p\mathrm{SO}(n)``
 of the point `p` on the [`Rotations`](@ref) manifold `M`
 with which the point `q` can be reached by the
-[`PolarRetraction`](https://juliamanifolds.github.io/ManifoldsBase.jl/stable/retractions.html#ManifoldsBase.PolarRetraction) from the point `p` after time 1.
+[`PolarRetraction`](@extref `ManifoldsBase.PolarRetraction`) from the point `p` after time 1.
 
 The formula reads
 ````math
@@ -191,7 +191,7 @@ inverse_retract(::Rotations, ::Any, ::Any, ::PolarInverseRetraction)
 
 Compute a vector from the tangent space ``T_p\mathrm{SO}(n)`` of the point `p` on the
 [`Rotations`](@ref) manifold `M` with which the point `q` can be reached by the
-[`QRRetraction`](https://juliamanifolds.github.io/ManifoldsBase.jl/stable/retractions.html#ManifoldsBase.QRRetraction) from the point `q` after time 1.
+[`QRRetraction`](@extref `ManifoldsBase.QRRetraction`) from the point `q` after time 1.
 """
 inverse_retract(::Rotations, ::Any, ::Any, ::QRInverseRetraction)