+++ src/manifolds/Stiefel.jl
@@ -78,7 +78,7 @@ end
     check_point(M::Stiefel, p; kwargs...)
 
 Check whether `p` is a valid point on the [`Stiefel`](@ref) `M`=``\operatorname{St}(n,k)``, i.e. that it has the right
-[`AbstractNumbers`](https://juliamanifolds.github.io/ManifoldsBase.jl/stable/types.html#number-system) type and ``p^{\mathrm{H}}p`` is (approximately) the identity, where ``⋅^{\mathrm{H}}`` is the
+[`AbstractNumbers`](@extref ManifoldsBase number-system) type and ``p^{\mathrm{H}}p`` is (approximately) the identity, where ``⋅^{\mathrm{H}}`` is the
 complex conjugate transpose. The settings for approximately can be set with `kwargs...`.
 """
 function check_point(M::Stiefel, p; kwargs...)
@@ -98,7 +98,7 @@ end
     check_vector(M::Stiefel, p, X; kwargs...)
 
 Checks whether `X` is a valid tangent vector at `p` on the [`Stiefel`](@ref)
-`M`=``\operatorname{St}(n,k)``, i.e. the [`AbstractNumbers`](https://juliamanifolds.github.io/ManifoldsBase.jl/stable/types.html#number-system) fits and
+`M`=``\operatorname{St}(n,k)``, i.e. the [`AbstractNumbers`](@extref ManifoldsBase number-system) fits and
 it (approximately) holds that ``p^{\mathrm{H}}X + \overline{X^{\mathrm{H}}p} = 0``,
 where ``⋅^{\mathrm{H}}`` denotes the Hermitian and ``\overline{⋅}`` the (elementwise) complex conjugate.
 The settings for approximately can be set with `kwargs...`.
@@ -119,7 +119,7 @@ end
 """
     default_inverse_retraction_method(M::Stiefel)
 
-Return [`PolarInverseRetraction`](https://juliamanifolds.github.io/ManifoldsBase.jl/stable/retractions.html#ManifoldsBase.PolarInverseRetraction) as the default inverse retraction for the
+Return [`PolarInverseRetraction`](@extref `ManifoldsBase.PolarInverseRetraction`) as the default inverse retraction for the
 [`Stiefel`](@ref) manifold.
 """
 default_inverse_retraction_method(::Stiefel) = PolarInverseRetraction()
@@ -127,14 +127,14 @@ default_inverse_retraction_method(::Stiefel) = PolarInverseRetraction()
 """
     default_retraction_method(M::Stiefel)
 
-Return [`PolarRetraction`](https://juliamanifolds.github.io/ManifoldsBase.jl/stable/retractions.html#ManifoldsBase.PolarRetraction) as the default retraction for the [`Stiefel`](@ref) manifold.
+Return [`PolarRetraction`](@extref `ManifoldsBase.PolarRetraction`) as the default retraction for the [`Stiefel`](@ref) manifold.
 """
 default_retraction_method(::Stiefel) = PolarRetraction()
 
 """
     default_vector_transport_method(M::Stiefel)
 
-Return the [`DifferentiatedRetractionVectorTransport`](https://juliamanifolds.github.io/ManifoldsBase.jl/stable/vector_transports.html#ManifoldsBase.DifferentiatedRetractionVectorTransport) of the [`PolarRetraction`]([`PolarRetraction`](https://juliamanifolds.github.io/ManifoldsBase.jl/stable/retractions.html#ManifoldsBase.PolarRetraction)
+Return the [`DifferentiatedRetractionVectorTransport`](@extref `ManifoldsBase.DifferentiatedRetractionVectorTransport`) of the [`PolarRetraction`]([`PolarRetraction`](@extref `ManifoldsBase.PolarRetraction`)
 as the default vector transport method for the [`Stiefel`](@ref) manifold.
 """
 function default_vector_transport_method(::Stiefel)
@@ -435,7 +435,7 @@ retract(::Stiefel, ::Any, ::Any, ::PadeRetraction)
 @doc raw"""
     retract(M::Stiefel, p, X, ::PolarRetraction)
 
-Compute the SVD-based retraction [`PolarRetraction`](https://juliamanifolds.github.io/ManifoldsBase.jl/stable/retractions.html#ManifoldsBase.PolarRetraction) on the
+Compute the SVD-based retraction [`PolarRetraction`](@extref `ManifoldsBase.PolarRetraction`) on the
 [`Stiefel`](@ref) manifold `M`. With ``USV = p + X`` the retraction reads
 
 ````math
@@ -447,7 +447,7 @@ retract(::Stiefel, ::Any, ::Any, ::PolarRetraction)
 @doc raw"""
     retract(M::Stiefel, p, X, ::QRRetraction)
 
-Compute the QR-based retraction [`QRRetraction`](https://juliamanifolds.github.io/ManifoldsBase.jl/stable/retractions.html#ManifoldsBase.QRRetraction) on the
+Compute the QR-based retraction [`QRRetraction`](@extref `ManifoldsBase.QRRetraction`) on the
 [`Stiefel`](@ref) manifold `M`. With ``QR = p + X`` the retraction reads
 
 ````math
@@ -546,7 +546,7 @@ end
 @doc raw"""
     vector_transport_direction(::Stiefel, p, X, d, ::DifferentiatedRetractionVectorTransport{CayleyRetraction})
 
-Compute the vector transport given by the differentiated retraction of the [`CayleyRetraction`](https://juliamanifolds.github.io/ManifoldsBase.jl/stable/retractions.html#ManifoldsBase.CayleyRetraction), cf. [Zhu:2016](@cite) Equation (17).
+Compute the vector transport given by the differentiated retraction of the [`CayleyRetraction`](@extref `ManifoldsBase.CayleyRetraction`), cf. [Zhu:2016](@cite) Equation (17).
 
 The formula reads
 ````math
@@ -561,7 +561,7 @@ with
 ````
 
 Since this is the differentiated retraction as a vector transport, the result will be in the
-tangent space at ``q=\operatorname{retr}_p(d)`` using the [`CayleyRetraction`](https://juliamanifolds.github.io/ManifoldsBase.jl/stable/retractions.html#ManifoldsBase.CayleyRetraction).
+tangent space at ``q=\operatorname{retr}_p(d)`` using the [`CayleyRetraction`](@extref `ManifoldsBase.CayleyRetraction`).
 """
 vector_transport_direction(
     M::Stiefel,